requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.6
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
